feat(Reward): 실행 결과 포장 클래스 생성
