+++ src/test/java/ladder/domain/ResultTest.java
@@ -13,32 +13,9 @@ public class ResultTest {
     @ParameterizedTest
     @ValueSource(strings = {"꽝", "5000", "꽝", "3000"})
     @DisplayName("실행 결과를 받아서 LadderResult를 생성한다.")
-    void generateTest(String result) {
-        Assertions.assertDoesNotThrow(() -> new Result(result));
+    void generateTest(String reward) {
+        Assertions.assertDoesNotThrow(() -> new Result(new Reward(reward)));
     }
 
-    @ParameterizedTest
-    @ValueSource(strings = {"가", "#", "a", "a1", "1.1"})
-    @DisplayName("\"꽝\" 이외의 실행 결과가 정수가 아니면 예외를 던진다.")
-    void validateRunResultNumericExceptNoLuckTest(String runResult) {
-        assertThatThrownBy(() -> new Result(runResult))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] \"꽝\" 이외의 실행 결과는 정수여야합니다.");
-    }
 
-    @Test
-    @DisplayName("실행 결과가 \"꽝\"이면 정수가 아니어도 예외가 발생하지 않는다.")
-    void validatePassWhenNoLuckTest() {
-        String runResult = "꽝";
-        Assertions.assertDoesNotThrow(() -> new Result(runResult));
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"0", "-1", "10001", "10002"})
-    @DisplayName("실행 결과가 정수일 때 1 이상 10000 이하가 아니면 예외를 던진다.")
-    void validateRunResultRangeTest(String runResult) {
-        assertThatThrownBy(() -> new Result(runResult))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] \"꽝\" 이외의 실행 결과는 1 이상 10000 이하의 정수여야합니다.");
-    }
 }
