+++ src/test/java/ladder/domain/ResultsTest.java
@@ -13,8 +13,8 @@ class ResultsTest {
     @Test
     @DisplayName("실행 결과들을 생성한다.")
     void generateTest() {
-        Result result1 = new Result("꽝");
-        Result result2 = new Result("5000");
+        Result result1 = new Result(new Reward("꽝"));
+        Result result2 = new Result(new Reward("5000"));
 
         Assertions.assertDoesNotThrow(() -> new Results(List.of(result1, result2)));
     }
